requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
qrcode==7.4.2
Pillow==10.0.1
Werkzeug==2.3.7
